Source Code of the Flash Torch app for Android developed be Ramandeep Singh Bakshi.
This is the source code to the current version of the Flash Torch app
for Android smartphones developed by Ramandeep Singh Bakshi.

At the time of release of this source code Flash Torch is at version 1.2
in the Google Play Store and it is this version that this source code
corresponds to.

To checkout the Flash Torch app by Ramandeep Singh Bakshi, visit this
link -
https://play.google.com/store/apps/details?id=com.ramandeepbakshi.flashtorch